# execution: reorganize function operators