--- a/execution/execution.go
+++ b/execution/execution.go
@@ -22,23 +22,19 @@ import (
 	"sort"
 	"time"
 
-	"github.com/prometheus/prometheus/promql"
-
-	"github.com/thanos-io/promql-engine/execution/noop"
-	"github.com/thanos-io/promql-engine/execution/remote"
-
 	"github.com/efficientgo/core/errors"
-
-	"github.com/prometheus/prometheus/storage"
-
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/storage"
 
 	"github.com/thanos-io/promql-engine/execution/aggregate"
 	"github.com/thanos-io/promql-engine/execution/binary"
 	"github.com/thanos-io/promql-engine/execution/exchange"
 	"github.com/thanos-io/promql-engine/execution/function"
 	"github.com/thanos-io/promql-engine/execution/model"
+	"github.com/thanos-io/promql-engine/execution/noop"
 	"github.com/thanos-io/promql-engine/execution/parse"
+	"github.com/thanos-io/promql-engine/execution/remote"
 	"github.com/thanos-io/promql-engine/execution/scan"
 	"github.com/thanos-io/promql-engine/execution/step_invariant"
 	engstore "github.com/thanos-io/promql-engine/execution/storage"
@@ -95,85 +91,13 @@ func newOperator(expr parser.Expr, storage *engstore.SelectorPool, opts *query.O
 		hints.Func = e.Func.Name
 		hints.Grouping = nil
 		hints.By = false
-
-		if e.Func.Name == "histogram_quantile" {
-			nextOperators := make([]model.VectorOperator, len(e.Args))
-			for i := range e.Args {
-				next, err := newOperator(e.Args[i], storage, opts, hints)
-				if err != nil {
-					return nil, err
-				}
-				nextOperators[i] = next
-			}
-
-			return function.NewHistogramOperator(model.NewVectorPool(stepsBatch), e.Args, nextOperators, stepsBatch)
-		}
-
-		// TODO(saswatamcode): Tracked in https://github.com/thanos-io/promql-engine/issues/23
-		// Based on the category we can create an apt query plan.
-		call, err := function.NewFunctionCall(e.Func)
-		if err != nil {
-			return nil, err
-		}
-
-		// TODO(saswatamcode): Range vector result might need new operator
-		// before it can be non-nested. https://github.com/thanos-io/promql-engine/issues/39
 		for i := range e.Args {
 			switch t := e.Args[i].(type) {
 			case *parser.MatrixSelector:
-				if call == nil {
-					return nil, parse.ErrNotImplemented
-				}
-
-				vs, filters, err := unpackVectorSelector(t)
-				if err != nil {
-					return nil, err
-				}
-
-				milliSecondRange := t.Range.Milliseconds()
-				if function.IsExtFunction(hints.Func) {
-					milliSecondRange += opts.ExtLookbackDelta.Milliseconds()
-				}
-
-				start, end := getTimeRangesForVectorSelector(vs, opts, milliSecondRange)
-				hints.Start = start
-				hints.End = end
-				hints.Range = milliSecondRange
-				filter := storage.GetFilteredSelector(start, end, opts.Step.Milliseconds(), vs.LabelMatchers, filters, hints)
-
-				numShards := runtime.GOMAXPROCS(0) / 2
-				if numShards < 1 {
-					numShards = 1
-				}
-
-				operators := make([]model.VectorOperator, 0, numShards)
-				for i := 0; i < numShards; i++ {
-					operator := exchange.NewConcurrent(
-						scan.NewMatrixSelector(model.NewVectorPool(stepsBatch), filter, call, e, opts, t.Range, vs.Offset, i, numShards),
-						2,
-					)
-					operators = append(operators, operator)
-				}
-
-				return exchange.NewCoalesce(model.NewVectorPool(stepsBatch), operators...), nil
+				return newRangeVectorFunction(e, t, storage, opts, hints)
 			}
 		}
-
-		// Does not have matrix arg so create functionOperator normally.
-		nextOperators := make([]model.VectorOperator, 0, len(e.Args))
-		for i := range e.Args {
-			// Strings don't need an operator
-			if e.Args[i].Type() == parser.ValueTypeString {
-				continue
-			}
-			next, err := newOperator(e.Args[i], storage, opts, hints)
-			if err != nil {
-				return nil, err
-			}
-			nextOperators = append(nextOperators, next)
-		}
-
-		return function.NewFunctionOperator(e, call, nextOperators, stepsBatch, opts)
+		return newInstantVectorFunction(e, storage, opts, hints)
 
 	case *parser.AggregateExpr:
 		hints.Func = e.Op.String()
@@ -293,6 +217,59 @@ func unpackVectorSelector(t *parser.MatrixSelector) (*parser.VectorSelector, []*
 	default:
 		return nil, nil, parse.ErrNotSupportedExpr
 	}
+}
+
+func newRangeVectorFunction(e *parser.Call, t *parser.MatrixSelector, storage *engstore.SelectorPool, opts *query.Options, hints storage.SelectHints) (model.VectorOperator, error) {
+	// TODO(saswatamcode): Range vector result might need new operator
+	// before it can be non-nested. https://github.com/thanos-io/promql-engine/issues/39
+	vs, filters, err := unpackVectorSelector(t)
+	if err != nil {
+		return nil, err
+	}
+
+	milliSecondRange := t.Range.Milliseconds()
+	if parse.IsExtFunction(e.Func.Name) {
+		milliSecondRange += opts.ExtLookbackDelta.Milliseconds()
+	}
+
+	start, end := getTimeRangesForVectorSelector(vs, opts, milliSecondRange)
+	hints.Start = start
+	hints.End = end
+	hints.Range = milliSecondRange
+	filter := storage.GetFilteredSelector(start, end, opts.Step.Milliseconds(), vs.LabelMatchers, filters, hints)
+
+	numShards := runtime.GOMAXPROCS(0) / 2
+	if numShards < 1 {
+		numShards = 1
+	}
+
+	operators := make([]model.VectorOperator, 0, numShards)
+	for i := 0; i < numShards; i++ {
+		operator, err := scan.NewMatrixSelector(model.NewVectorPool(stepsBatch), filter, e, opts, t.Range, vs.Offset, i, numShards)
+		if err != nil {
+			return nil, err
+		}
+		operators = append(operators, exchange.NewConcurrent(operator, 2))
+	}
+
+	return exchange.NewCoalesce(model.NewVectorPool(stepsBatch), operators...), nil
+}
+
+func newInstantVectorFunction(e *parser.Call, storage *engstore.SelectorPool, opts *query.Options, hints storage.SelectHints) (model.VectorOperator, error) {
+	nextOperators := make([]model.VectorOperator, 0, len(e.Args))
+	for i := range e.Args {
+		// Strings don't need an operator
+		if e.Args[i].Type() == parser.ValueTypeString {
+			continue
+		}
+		next, err := newOperator(e.Args[i], storage, opts, hints)
+		if err != nil {
+			return nil, err
+		}
+		nextOperators = append(nextOperators, next)
+	}
+
+	return function.NewFunctionOperator(e, nextOperators, stepsBatch, opts)
 }
 
 func newShardedVectorSelector(selector engstore.SeriesSelector, opts *query.Options, offset time.Duration) (model.VectorOperator, error) {