--- a/execution/parse/functions.go
+++ b/execution/parse/functions.go
@@ -1,6 +1,12 @@
 package parse
 
-import "github.com/thanos-io/promql-engine/parser"
+import (
+	"fmt"
+
+	"github.com/efficientgo/core/errors"
+
+	"github.com/thanos-io/promql-engine/parser"
+)
 
 var Functions = map[string]*parser.Function{
 	"xdelta": {
@@ -18,4 +24,19 @@ var Functions = map[string]*parser.Function{
 		ArgTypes:   []parser.ValueType{parser.ValueTypeMatrix},
 		ReturnType: parser.ValueTypeVector,
 	},
+}
+
+// IsExtFunction is a convenience function to determine whether extended range calculations are required.
+func IsExtFunction(functionName string) bool {
+	_, ok := Functions[functionName]
+	return ok
+}
+
+func UnknownFunctionError(f *parser.Function) error {
+	msg := fmt.Sprintf("unknown function: %s", f.Name)
+	if _, ok := parser.Functions[f.Name]; ok {
+		return errors.Wrap(ErrNotImplemented, msg)
+	}
+
+	return errors.Wrap(ErrNotSupportedExpr, msg)
 }