… execution/scan/functions.go