* distinguish between functions taking an instant vector and functions taking a range vector
* since `scan` does not depend on `function` anymore we can hide a couple structs and functions
* some cleanup in execution.go